Update ast2markdown.xsl
Proposed fix for https://github.com/jelovirt/org.lwdita/issues/46 Nested Lists not properly generated in output Markdown content when list Item starts with bold, italics, codeph...
The examples in this issue demonstrate what this fix will do. Be aware that for all list item between "-" for unordered / "1." for ordered list and the follwoing text of the list item in the future exactly one blank " " will be rendered to prevent that some following content is interpreted as codeblock.